saving state

+++ src/main/java/matrix/SquareMatrix.java
@@ -2,10 +2,11 @@ package matrix;
 
 public interface SquareMatrix<T extends SquareMatrix<T>> extends Matrix<T> {
 
-  float determinant();
+  double determinant();
 
   T inverse();
 
   @Override
   T transpose();
+
 }
